implement & test image upload
