+++ src/DataFixtures/Model/AssetFile.php
@@ -11,7 +11,6 @@
 
 namespace App\DataFixtures\Model;
 
-use Symfony\Component\HttpFoundation\File\Exception\FileException;
 use Symfony\Component\HttpFoundation\File\UploadedFile;
 
 class AssetFile extends UploadedFile
@@ -19,22 +18,10 @@ class AssetFile extends UploadedFile
     public function __construct(string $path)
     {
         $originalName = pathinfo($path, PATHINFO_BASENAME);
-        parent::__construct($path, $originalName);
-    }
 
-    public function move(string $directory, string $name = null)
-    {
-        $target = $this->getTargetFile($directory, $name);
+        $newPath = $path.'.tmp';
+        copy($path, $newPath);
 
-        set_error_handler(function ($type, $msg) use (&$error) { $error = $msg; });
-        $renamed = copy($this->getPathname(), $target);
-        restore_error_handler();
-        if (!$renamed) {
-            throw new FileException(sprintf('Could not copy the file "%s" to "%s" (%s).', $this->getPathname(), $target, strip_tags($error)));
-        }
-
-        chmod($target, 0666 & ~umask());
-
-        return $target;
+        parent::__construct($newPath, $originalName);
     }
 }
