+++ tests/Api/ConstructionSiteTest.php
@@ -12,22 +12,19 @@
 namespace App\Tests\Api;
 
 use ApiPlatform\Core\Bridge\Symfony\Bundle\Test\ApiTestCase;
-use ApiPlatform\Core\Bridge\Symfony\Bundle\Test\Client;
-use App\DataFixtures\Model\AssetFile;
-use App\Entity\ConstructionSite;
 use App\Tests\DataFixtures\TestConstructionSiteFixtures;
 use App\Tests\DataFixtures\TestUserFixtures;
 use App\Tests\Traits\AssertApiTrait;
 use App\Tests\Traits\AuthenticationTrait;
-use Doctrine\Persistence\ManagerRegistry;
+use App\Tests\Traits\TestDataTrait;
 use Liip\TestFixturesBundle\Test\FixturesTrait;
-use Symfony\Component\HttpFoundation\Response as StatusCode;
 
 class ConstructionSiteTest extends ApiTestCase
 {
     use FixturesTrait;
     use AssertApiTrait;
     use AuthenticationTrait;
+    use TestDataTrait;
 
     public function testInvalidMethods()
     {
@@ -37,13 +34,6 @@ class ConstructionSiteTest extends ApiTestCase
 
         $testConstructionSite = $this->getTestConstructionSite();
         $this->assertApiOperationUnsupported($client, '/api/construction_sites/'.$testConstructionSite->getId(), 'DELETE', 'PUT', 'PATCH');
-    }
-
-    private function getTestConstructionSite(): ConstructionSite
-    {
-        $constructionSiteRepository = static::$container->get(ManagerRegistry::class)->getRepository(ConstructionSite::class);
-
-        return $constructionSiteRepository->findOneByName(TestConstructionSiteFixtures::TEST_CONSTRUCTION_SITE_NAME);
     }
 
     public function testValidMethodsNeedAuthentication()
@@ -87,32 +77,5 @@ class ConstructionSiteTest extends ApiTestCase
 
         $this->assertApiPostFieldsRequired($client, '/api/construction_sites', $sample);
         $this->assertApiPostFieldsPersisted($client, '/api/construction_sites', $sample);
-    }
-
-    public function testPostImage()
-    {
-        $client = $this->createClient();
-        $this->loadFixtures([TestUserFixtures::class, TestConstructionSiteFixtures::class]);
-        $this->loginApiTestUser($client);
-
-        $testConstructionSite = $this->getTestConstructionSite();
-
-        $uploadedFile = new AssetFile(__DIR__.'/../../assets/samples/Test/preview_2.jpg');
-        $this->assertApiPostUploadFile($client, '/api/construction_sites/'.$testConstructionSite->getId().'/image', $uploadedFile);
-    }
-
-    private function assertApiPostUploadFile(Client $client, string $url, AssetFile $file)
-    {
-        $response = $client->getKernelBrowser()->request('POST', $url,
-            ['headers' => ['Content-Type' => 'application/json']],
-            ['file' => $file]
-        );
-
-        $this->assertResponseStatusCodeSame(StatusCode::HTTP_OK);
-
-        $content = $response->text();
-        $hydraPayload = json_decode($content, true);
-
-        dump($hydraPayload);
     }
 }
